pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "cloud-posture-watch"
version = "0.1.0"
description = "Cloud security posture assessment for AWS, Azure, and GCP"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [{ name = "Hiago Kin Levi" }]

# Runtime dependencies
dependencies = [
    # CLI framework
    "click>=8.1",
    # Data validation and settings management
    "pydantic>=2.6",
    # AWS SDK
    "boto3>=1.34",
    # Azure identity / authentication
    "azure-identity>=1.16",
    # Azure storage management (account settings, encryption, logging)
    "azure-mgmt-storage>=21.0",
    # Azure network management (NSG rules)
    "azure-mgmt-network>=25.0",
    # GCP storage client
    "google-cloud-storage>=2.16",
    # Structured logging
    "structlog>=24.1",
    # Table formatting for CLI output
    "tabulate>=0.9",
    # Jinja2 templating for report generation
    "jinja2>=3.1",
    # YAML parsing for baseline files
    "pyyaml>=6.0",
    # .env file support
    "python-dotenv>=1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "black>=24.0",
    "ruff>=0.4",
    "mypy>=1.10",
    # Type stubs for boto3 (improves IDE experience)
    "boto3-stubs[s3,cloudtrail,ec2]>=1.34",
]

[project.scripts]
# Main CLI entry point
k1n-posture = "cli.main:cli"

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP"]
ignore = ["E501"]  # Line length handled by black

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "--cov=. --cov-report=term-missing"

[tool.mypy]
python_version = "3.11"
strict = true
ignore_missing_imports = true
